move `Route` convenient logic to their own files - remove convenience using `MediaType` (alternatively use the `some MediaTypeProtocol` convenience logic) - `String` now conforms to `HTTPRequestMethodProtocol` - use `String` for the convenient `Route` methods (prevously required the http request method package traits)
