+++ Sources/DestinyBlueprint/http/HTTPRequestMethodProtocol.swift
@@ -2,3 +2,11 @@
 public protocol HTTPRequestMethodProtocol: Sendable, ~Copyable {
     func rawNameString() -> String
 }
+
+extension String: HTTPRequestMethodProtocol { // TODO: use under a package trait
+    @inlinable
+    @inline(__always)
+    public func rawNameString() -> String {
+        self
+    }
+}
